requirements.txt
websockets>=12.0
aiohttp
asyncio
requests
maim_message
loguru
pillow
tomli
discord.py>=2.3.2